% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLCAres.R
\docType{data}
\name{BLCAres}
\alias{BLCAres}
\title{Trimmed median LRR values for autosomal and male-specific Y (msY, chrY:2694521-59034049, hg18/GRCh36) regions from 595 TCGA BLCA male samples.}
\format{A MADloy-class object with:
\describe{
  \item{Target}{Median LRR for the target region}
  \item{Ref.region}{Median LRR for the autosomal chromosomes}
  \item{par}{parameters specifying the regions analyzed}
  ...
}}
\usage{
BLCAres
}
\description{
A MADloy object containing the median LRR values for the autosomal chromosomes and chromosome Y from position 2,694,521 to 59,034,049 from 1038 TCGA BLCA male samples.
}
\keyword{datasets}
