% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSex.R
\name{checkSex}
\alias{checkSex}
\title{Check sex for SNP array samples}
\usage{
checkSex(files, rsCol = 1, ChrCol = 2, PosCol = 3, LRRCol = 4, mc.cores,
  trim = 0.2, quiet = FALSE)
}
\arguments{
\item{files}{A single file path (APT platform and MAD platform), a vector of 
file paths (MAD platform) or a MAD rawData folder path containing files 
ready to be processed with MAD (MAD platform).}

\item{rsCol}{The position of the column with the SNP identifier.}

\item{ChrCol}{The position of the column with the Chromosome field.}

\item{PosCol}{The position of the column with the Position field.}

\item{LRRCol}{The position of the column with the LRR identifier.}

\item{mc.cores}{The number of cores used to perform the function. By default 
is set to 1.}

\item{trim}{the fraction (0 to 0.5) of probes to be trimmed when summaryzing LRR. 
This argument tries to control the effect of having CNVs across genome. Default is 0.2.}

\item{quiet}{Should the function not inform about the status of the process. 
By default is FALSE.}
}
\value{
A data.frame object that summarizes the LRR in chrX and chrY in
  the analyzed samples.
}
\description{
Check sex for SNP array samples
}
\examples{
\dontrun{
checkSex(filepath, mc.cores=20)}
}
