% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLOY.R
\name{getLOY}
\alias{getLOY}
\title{Detection algorithm to detect Loss of Y events in MADloy or MADseqLOY data}
\usage{
getLOY(object, ref = "22", k, cutoff = 0.9, arbvar = FALSE, ...)
}
\arguments{
\item{object}{A MADloy or MADseqLOY object.}

\item{ref}{One of the reference chromosomes used in the \code{MADloy} or \code{MADseqLOY}
functions.}

\item{k}{Number of expected classification groups}

\item{...}{Other parameters}
}
\value{
An object of "LOY" class that summarizes the LOY events detected in
  the analyzed samples
}
\description{
Detection algorithm to detect Loss of Y events in MADloy or MADseqLOY data
}
\examples{
\dontrun{
getLOY(resMADseqLOY)
getLOY(resMADloy)}
}

