% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMedian.R
\name{getMedian}
\alias{getMedian}
\title{Get median values from MADseqLOY or MADloy objects}
\usage{
getMedian(object)
}
\arguments{
\item{object}{A MADloy or MADseqLOY object from the \code{MADloy} or
  \code{MADseqLOY} functions.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function retrieves the median values from the MADloy or MADseqLOY
objects and prepares the data to be analyzed for LOY events.
}
\examples{
\dontrun{
getMedian(resMADloy)
getMedian(resMADseqLOY)}
}

