% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProb.R
\name{getProb}
\alias{getProb}
\title{Get probability matrix of three status (LOY, normal, XYY)}
\usage{
getProb(x)
}
\arguments{
\item{x}{an object of class 'LOY'}
}
\value{
a matrix of probabilities
}
\description{
Get probability matrix of three status (LOY, normal, XYY)
}
