% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamples.R
\name{getSamples}
\alias{getSamples}
\title{Return the samples with the chosen classification}
\usage{
getSamples(x, type = "LOY")
}
\arguments{
\item{x}{A MADloy, LOY or MADloyBdev object.}

\item{type}{A string with the chosen classification.}
}
\value{
A vector of the samples with the chosen classification.
}
\description{
Return the samples with the chosen classification
}
\examples{
\dontrun{
getSamples(resMADloy)}
}
