% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Get summary values from MADseqLOY or MADloy objects}
\usage{
getSummary(object)
}
\arguments{
\item{object}{A MADloy or MADseqLOY object from the \code{MADloy} or 
\code{MADseqLOY} functions.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function retrieves the summary values from the MADloy or MADseqLOY
objects and prepares the data to be analyzed for LOY events.
}
\examples{
\dontrun{
getSummary(resMADloy)
getSummary(resMADseqLOY)}
}

