% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MADloyDataAll.R
\docType{data}
\name{MADloyDataAll}
\alias{MADloyDataAll}
\title{Median chr21, chr22 and chrY:7000000-30000000 LRR values from 87 HapMap 3 samples.}
\format{A MADloy-class object with:
\describe{
  \item{Target}{Median LRR for the target region}
  \item{Ref.region.1}{Median LRR for the reference region 1}
  \item{Ref.region.2}{Median LRR for the reference region 2}
  \item{par}{parameters specifying the regions analyzed}
  ...
}}
\usage{
MADloyDataAll
}
\description{
A dataset containing the median LRR values for the whole chromosome 21, whole chromosome 22 and chromosome Y from position 7,000,000 to 30,000,000 from 87 HapMap 3 samples.
}
\keyword{datasets}

