% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MADseqLOY.R
\name{MADseqLOY}
\alias{MADseqLOY}
\title{Check the loss of Y events in a given folder of .bam files}
\usage{
MADseqLOY(files, exomeTargets, target.region = "chrY:7000000-30000000",
  ref.region.1 = "chr21", ref.region.2 = "chr22", mc.cores, quiet = FALSE,
  skip = 2, ...)
}
\arguments{
\item{files}{A folder path where the .bam files are or a vector of file
paths. This function searches in recursive folders.}

\item{exomeTargets}{A file path to a .bed file with the regions targeted by
the sequencing method used.}

\item{target.region}{The chromosome or region to be compared with the other
regions. By default is the region chrY:7000000-3000000 but it can be
changed.}

\item{ref.region.1}{First chromosome or region to be compared with the Y
region in UCSC style (i.e. "chr21" or "chr21:1000-10000").}

\item{ref.region.2}{Second chromosome or region to be compared with the Y
region in UCSC style (i.e. "chr22" or "chr22:1000-10000").}

\item{mc.cores}{umber of cores to be used with this function. If there are
more cores than samples, the number of cores will be limited to the number
of samples. By default set to 1.}

\item{quiet}{Should the function not inform about the status of the process.
By default is FALSE.}

\item{skip}{Number of lines to be skipped in the targets file if necessary.
By default is set to 1.}

\item{...}{Other parameters of the read.targets function.}
}
\value{
The returned value is an object of "MADseqLOY" class with four lists,
  one regarding the target region coverage called "target", two regarding the
  reference regions to be compared against called "reference.1" and
  "reference.2" and finally one with the parameters used in the query.
}
\description{
MADseqLOY check all the bams in a given folder, by using the target regions
of the technology used to sequence the sample. The coverage is computed for
the region chrY:7000000-3000000 and another given by the query argument
}
\examples{
\dontrun{
MADseqLOY(files=bamFile, reference=targetFile, skip=0)}
}

