% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndLRR.R
\name{plotIndLRR}
\alias{plotIndLRR}
\title{Plots the SNP LRR data of an individual}
\usage{
plotIndLRR(x, sample, rsCol = 1, ChrCol = 2, PosCol = 3, LRRCol = 4,
  ...)
}
\arguments{
\item{x}{A MADloy object from the \code{MADloy} functions.}

\item{sample}{The identifier of a sample in the MADloy object.}

\item{rsCol}{Column of the MAD file with the name of the SNP.}

\item{ChrCol}{Column of the MAD file with the chromosome information.}

\item{PosCol}{Column of the MAD file with the position information.}

\item{LRRCol}{Column of the MAD file with the LRR information.}

\item{...}{Other parameters of the generic plot function.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function draws the LRR data for an individual
}
\examples{
\dontrun{
plotindLRR(resMADloy, 'SAMPLE')}
}
