% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndSNP.R
\name{plotIndSNP}
\alias{plotIndSNP}
\title{Plots the SNP data of chromosome Y for an individual}
\usage{
plotIndSNP(x, sample, rsCol = 1, ChrCol = 2, PosCol = 3,
  LRRCol = 4, BAFCol = 5, offset = 0, scaleY = FALSE, ...)
}
\arguments{
\item{x}{A MADloy object from the \code{MADloy} functions.}

\item{sample}{The identifier of a sample in the MADloy object.}

\item{rsCol}{Column of the MAD file with the name of the SNP.}

\item{ChrCol}{Column of the MAD file with the chromosome information.}

\item{PosCol}{Column of the MAD file with the position information.}

\item{LRRCol}{Column of the MAD file with the LRR information.}

\item{BAFCol}{Column of the MAD file with the BAF information.}

\item{offset}{offset value to adjust msY region LRR. By default is set to 0.}

\item{scaleY}{Should the LRR in msY region be scaled to simulate a diploid region - ploidy(msY)+1. By default is set to TRUE.}

\item{...}{Any other graphical parameter.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function draws the LRR and BAF data of chromosome Y for an individual analyzed with madloy or checkBdev
}
\examples{
\dontrun{
plotindSNP(resMADloy, "SAMPLE")}
}
