% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNIG.R
\name{plotNIG}
\alias{plotNIG}
\title{Plot of the distribution of the LRR values for the different models}
\usage{
plotNIG(x, tit)
}
\arguments{
\item{x}{A MADloy object from the \code{madloy} function.}

\item{tit}{Title of the plot.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function draws the LRR values distribution for the Normal, Normal Inverse Gaussian (NIG) and Empiric Cumulative Distribution functions from a LOY object.
}
\examples{
\dontrun{
plotNIG(resMADloy)}
}
