% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRADresBdev.R
\docType{data}
\name{PRADresBdev}
\alias{PRADresBdev}
\title{Trimmed median LRR values and Bdev values for PAR1, PAR2, p arm and q arm of chromosome Y from 242 TCGA PRAD male samples with putative alterated values.}
\format{A MADloyBdev-class object with:
\describe{
  \item{cl}{Classification}
  \item{Ref.region}{Median LRR for the autosomal chromosomes}
  \item{par}{parameters specifying the regions analyzed}
  ...
}}
\usage{
PRADresBdev
}
\description{
Trimmed median LRR values and Bdev values for PAR1, PAR2, p arm and q arm of chromosome Y from 242 TCGA PRAD male samples with putative alterated values.
}
\keyword{datasets}

