% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBdevMAD.R
\name{processBdevMAD}
\alias{processBdevMAD}
\title{Process MAD file to obtain Bdev}
\usage{
processBdevMAD(file, query, rsCol, ChrCol, PosCol, BAFCol, top, bot)
}
\arguments{
\item{file}{File path of the MAD file to be processed.}

\item{query}{Region of interest to check the LRR in rangedData format.}

\item{rsCol}{Column of the MAD file with the name of the SNP.}

\item{ChrCol}{Column of the MAD file with the chromosome information.}

\item{PosCol}{Column of the MAD file with the position information.}

\item{BAFCol}{Column of the MAD file with the B Allele Frequency information.}

\item{top}{Superior treshold to consider an heterozygous allele.}

\item{bot}{Inferior treshold to consider an heterozygous allele.}
}
\value{
A list with the Bdev for the query value.
}
\description{
Process a MAD file to retrieve the B deviation in a region of
interest.
}
\examples{
\dontrun{
processBdevMAD(file=file.path, query=rangedDataROI, rsCol=1, chrCol=2, PosCol=3, LRRCol=4)}
}

