% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processMAD.R
\name{processMAD}
\alias{processMAD}
\title{Process MAD file}
\usage{
processMAD(file, query, rsCol, ChrCol, PosCol, LRRCol)
}
\arguments{
\item{file}{File path of the MAD file to be processed.}

\item{query}{Region of interest to check the LRR in rangedData format.}

\item{rsCol}{Column of the MAD file with the name of the SNP.}

\item{ChrCol}{Column of the MAD file with the chromosome information.}

\item{PosCol}{Column of the MAD file with the position information.}

\item{LRRCol}{Column of the MAD file with the LRR information.}
}
\value{
A list with the mean of the LRR for the query value.
}
\description{
Process a MAD file to retrieve the median log R ratio of a region of
interest.
}
\examples{
\dontrun{
processMAD(file=file.path, query=rangedDataROI, rsCol=2, chrCol=3, PosCol=4, LRRCol=5)}
}

