% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBdev.R
\name{checkBdev}
\alias{checkBdev}
\title{Check the B deviation values in PAR1 and PAR2 regions to check possible Loss of Y events in a folder or files.}
\usage{
checkBdev(files, rsCol = 1, ChrCol = 2, PosCol = 3, BAFCol = 4,
  top = 0.9, bot = 0.1, mc.cores, quiet = FALSE, hg = "hg18", ...)
}
\arguments{
\item{files}{A single file path (APT platform and MAD platform), a vector of 
file paths (MAD platform) or a MAD rawData folder path containing files 
ready to be processed with MAD (MAD platform).}

\item{rsCol}{The position of the column with the SNP identifier.}

\item{ChrCol}{The position of the column with the Chromosome field.}

\item{PosCol}{The position of the column with the Position field.}

\item{BAFCol}{The position of the column with the BAF identifier.}

\item{top}{Superior treshold to consider an heterozygous allele.}

\item{bot}{Inferior treshold to consider an heterozygous allele.}

\item{mc.cores}{The number of cores used to perform the function. By default 
is set to 1.}

\item{quiet}{Should the function not inform about the status of the process. 
By default is FALSE.}

\item{...}{Other parameters}
}
\value{
A MADloyBdev object that contains the Bdev values for the two PAR regions for all the files 
  analyzed.
}
\description{
checkBdev checks the B deviation values of all the MAD files specified or in
a path to check for possible Loss of Y events. The Bdev is computed by default for the 
PAR1 and PAR2 regions.
}
\examples{
\dontrun{
checkBdev(filepath, mc.cores=2)}
}
\seealso{
\code{\link{getLOY}} to process results from \code{MADloy}
}

