% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLOY.R
\name{getLOY}
\alias{getLOY}
\title{Detection algorithm to detect Loss of Y events in MADloy or MADseqLOY data}
\usage{
getLOY(object, offset, pval.sig, k, cutoff, ...)
}
\arguments{
\item{object}{A MADloy or MADseqLOY object.}

\item{offset}{Offset value of SNP array data in the LRR of chromosome Y. That is, value to
guarantee that mean LRR at chrmosome Y is centered at 0.  Default 0 since LRR at 
m-LRR region is expected to be centered at O.}

\item{pval.sig}{pval.sig p-value treshold to be used in the classification test.}

\item{k}{Number of groups. Only necessary in NGS data.}

\item{cutoff}{cutoff value. Only necessari in NGS data.}

\item{...}{Other parameters.}
}
\value{
An object of class 'LOY' that summarizes the LOY events detected in
  the analyzed samples
}
\description{
Detection algorithm to detect Loss of Y events in MADloy or MADseqLOY data
}
\examples{
\dontrun{
getLOY(resMADseqLOY)
getLOY(resMADloy)}
}
