% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLOY.R
\name{getLOY}
\alias{getLOY}
\title{Detection algorithm to detect Loss of Y events in MADloy data}
\usage{
getLOY(object, coef = 3, ...)
}
\arguments{
\item{object}{A MADloy object.}

\item{coef}{this determines how far the outliers are considered.}

\item{...}{Other parameters.}
}
\value{
An object of class 'LOY' that summarizes the LOY events detected in
  the analyzed samples
}
\description{
Detection algorithm to detect Loss of Y events in MADloy data
}
\examples{
\dontrun{
getLOY(resMADloy)}
}
