% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Get summary values from a MADloy object}
\usage{
getSummary(object)
}
\arguments{
\item{object}{A MADloy object from the \code{MADloy} function.}
}
\value{
A data table with the results for all samples in columns.
}
\description{
This function retrieves the summary values from a MADloy 
object and prepares the data to be analyzed for LOY events.
}
\examples{
\dontrun{
getSummary(resMADloy)}
}
