% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MADloy.R
\name{MADloy}
\alias{MADloy}
\title{Check the mean LRR values to detect Loss of Y events in a folder or files.}
\usage{
MADloy(files, target.region = "chrY:7000000-30000000",
  ref.region.1 = "chr21", ref.region.2 = "chr22", rsCol = 1, ChrCol = 2,
  PosCol = 3, LRRCol = 4, mc.cores, quiet = FALSE, ...)
}
\arguments{
\item{files}{A single file path (APT platform and MAD platform), a vector of
file paths (MAD platform) or a MAD rawData folder path containing files
ready to be processed with MAD (MAD platform).}

\item{target.region}{The chromosome or region to be compared with the other
regions. By default is the region chrY:7000000-3000000 but it can be
changed.}

\item{ref.region.1}{First chromosome or region to be compared with the Y
region in UCSC style (i.e. "chr21" or "chr21:1000-10000").}

\item{ref.region.2}{Second chromosome or region to be compared with the Y
region in UCSC style (i.e. "chr22" or "chr22:1000-10000").}

\item{rsCol}{The position of the column with the SNP identifier.}

\item{ChrCol}{The position of the column with the Chromosome field.}

\item{PosCol}{The position of the column with the Position field.}

\item{LRRCol}{The position of the column with the LRR identifier.}

\item{mc.cores}{The number of cores used to perform the function. By default
is set to 1.}

\item{quiet}{Should the function not inform about the status of the process.
By default is FALSE.}

\item{...}{Other parameters}
}
\value{
A MADloy object that contains the LRR means for all the files
  analyzed.
}
\description{
MADloy check the median log R ratio(LRR) of all the MAD files specified or in
a path to detect Loss of Y events. The LRR is computed by default for the
regions chr21, chr22 and chrY:7000000-3000000.
}
\examples{
\dontrun{
MADloy(filepath, mc.cores=2)}
}
\seealso{
\code{\link{getLOY}} to process results from \code{MADloy}
}

