% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MADloy.R
\name{plot.MADloy}
\alias{plot.MADloy}
\title{Plot of MADloy object}
\usage{
\method{plot}{MADloy}(x, labels, print.labels = FALSE, threshold = -0.6,
  cex.label = 0.8, ...)
}
\arguments{
\item{print.labels}{should sample labels be showed}

\item{threshold}{threshold to draw sample labels}

\item{cex.label}{size of the labels}
}
\description{
This function plots the summarized LRR of a set of samples
}
