% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndSNP.R
\name{plotIndSNP}
\alias{plotIndSNP}
\title{Plots the SNP data of an individual}
\usage{
plotIndSNP(x, sample, rsCol = 1, ChrCol = 2, PosCol = 3, LRRCol = 4,
  BAFCol = 5, offset = 0, hg = "hg18", ...)
}
\arguments{
\item{x}{A MADloy object from the \code{MADloy} functions or A single file path (APT platform and MAD platform).}

\item{sample}{The identifier of a sample in the MADloy object when x is a MADloy object.}

\item{rsCol}{Column of the MAD file with the name of the SNP.}

\item{ChrCol}{Column of the MAD file with the chromosome information.}

\item{PosCol}{Column of the MAD file with the position information.}

\item{LRRCol}{Column of the MAD file with the LRR information.}

\item{BAFCol}{Column of the MAD file with the BAF information.}

\item{offset}{offset applied to the msY region.}

\item{hg}{Human genome build version. It can be "hg18", "hg19" or "GRCh38". Set by default to "hg18". Only used when a file path is given.}

\item{...}{Any other graphical parameter.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function draws the LRR and BAF data for an individual
}
\examples{
\dontrun{
plotindSNP(resMADloy, "SAMPLE")}
}
