% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNIG.R
\name{plotNIG}
\alias{plotNIG}
\title{Plot of LRR cummulative distribution function (CDF) of two models}
\usage{
plotNIG(object, tit)
}
\arguments{
\item{object}{An object of class MADloy from the \code{madloy} function.}

\item{tit}{Title of the plot.}
}
\value{
A data table with the results for all samples in columns
}
\description{
This function plots the CDF of the LRR values and the expected CDF under Normal and Inverse Gaussian (NIG) distributions.
}
\examples{
\dontrun{
plotNIG(resMADloy)}
}
